+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 5.8.0.0 - 2025-07-XX
 * Added support for exposure rejection for humidity.
-* TWILIGHT OFFSET?
+* Added support for an offset in minutes from twilight acceptable times.
 * EP DISABLE?
 
 ## 5.7.1.0 - 2025-06-28